% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpolateFunction.R
\name{interpolate}
\alias{interpolate}
\title{Data interpolation across soil depths}
\usage{
interpolate(Dataset, interCol, datRange)
}
\arguments{
\item{Dataset}{a dataframe of observed soil data}

\item{interCol}{the column over which the approx function embedded in interpolate() performs interpolation.}

\item{datRange}{The range of points to which the interCol column will be expanded.}
}
\value{
A dataframe of depth-interpolated values
}
\description{
Data interpolation across soil depths
}
\examples{
interpolate(InterP2, 19, 1:193)
}
