% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soil_Stats_function.R
\name{SoilCorMat}
\alias{SoilCorMat}
\title{Correlation of interpolated soil datasets}
\usage{
SoilCorMat(SoilDat, startCol, endCol)
}
\arguments{
\item{SoilDat}{A dataframe of soil observations for which the user would like to calculate correlations.}

\item{startCol}{Specifies dataframe column where correlation calculations should begin}

\item{endCol}{Specifies dataframe column for correlations calculations to end}
}
\value{
A correlation matrix of calculated r and p values.
}
\description{
\code{SoilCorMat()} is a function that takes a dataframe of soil observations and analyzes the correlations between each column of variables.
}
\section{Note}{
 this function requires the "Hmisc" package
}

\examples{
SoilCorMat(CO2, startCol = 4, endCol = 5)
}
